bug/IVYPORTAL-17514-Fix-GUI-tests-LTS
- Update resetFormattingLanguage method
